More ci tweaks, and additional CLI restructuring
