+++ packages/cli/jest.config.mjs
@@ -3,4 +3,8 @@ import base from '../../jest.config.base.mjs';
 export default {
   ...base,
   displayName: 'Gitbeaker CLI',
+  moduleNameMapper: {
+    "@gitbeaker/core/map.json": '<rootDir>/test/__mocks__/map.json',
+    ...base.moduleNameMapper,
+  }
 };
