+++ packages/cli/package.json
@@ -32,7 +32,7 @@
   ],
   "scripts": {
     "build": "tsup src/index.ts --format esm --treeshake --target node18",
-    "test:integration": "jest --maxWorkers=50% test/integration",
+    "test:unit": "jest --maxWorkers=50% test/unit",
     "test:e2e": "jest --maxWorkers=50% test/e2e",
     "format:docs": "prettier '**/(*.json|.yml|.js|.md)' --ignore-path ../../.prettierignore",
     "format:docs:fix": "yarn format:docs --write",
